feat: renames core module to client
