+++ client/client.go
@@ -1,4 +1,4 @@
-package core
+package client
 
 import (
 	"bytes"
